Add files via upload
This project aims to perform Customer Segmentation of USA Mall's customers, through the use of machine learning in the form of clustering (Python &  Jupyter).  This is the perfect way for a sizeable corporate or even a small firm, to identify marketing groups from a larger population directly through the data collected, to get the most accurate prediction of what kind of customers should be a business target, in the next marketing ad campaign. All this work is made by gathering information about the shopping groups based on income, age, and the mall shopping score.